+++ src/index.ts
@@ -1,7 +1,6 @@
 export { createQuery } from "./create-query";
 export { createBaseMutation } from "./create-base-mutation";
 export { createMutation } from "./create-mutation";
-export { cacheStore, CacheEntry } from "./cache-store";
 export type {
   QueryKey,
   QueryOptions,